requirements.txt
asgiref
backports.zoneinfo
certifi
charset-normalizer
click
Django
django-restframework
Flask
idna
itsdangerous
Jinja2
Jinja2
MarkupSafe
pycryptodome
requests
sqlparse
urllib3
Werkzeug
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-import-export==2.8.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.23.2
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
Pillow==9.2.0
PyAutoGUI==0.9.53
pyclick==0.0.2
pycryptodome==3.15.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.12
Werkzeug==2.2.2
xlib==0.21
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
